refactored degator end to delegating.py and corresponding tests and used credentialing design for IdentifierResource end usage
